requirements.txt
aiohttp
mnemonic
bip32utils
tronpy
python-telegram-bot
requests